pyproject.toml
[tool.poetry]
name = "isbnchile-ingestion"
version = "0.1.0"
description = "Ingesta códigos ISBN"
authors = ["Jose Pedro <peterotool@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.5"
numpy = "^1.26.4"
pandas = "^2.2.2"
scrapy = "^2.11.2"
python-dotenv = "^1.0.1"


[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

line-length = 120
indent-width = 4

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.12.
target-version = "py312"

[tool.ruff.lint]

# Enable pycodestyle (E) , Pyflakes (F), isort (I) an pyupgrade (UP)  codes by default.
select = ["E", "F", "I", "UP"]

#  Never enforce Line too long ({width} > {limit} (E501)
ignore = ["E501","F401"]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
# but never try to fix `F401` (unused imports).
unfixable = ["F401"]


[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

[tool.poetry.group.dev.dependencies]
ruff = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
